update from mac

+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: ohamadou <ohamadou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 01:23:05 by ohamadou          #+#    #+#             */
-/*   Updated: 2022/11/15 19:22:55 by ohamadou         ###   ########.fr       */
+/*   Updated: 2022/11/25 18:40:06 by ohamadou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 void	*ft_memset(void *ptr, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
